+++ src/engine/physics/ICollider.java
@@ -1,11 +1,12 @@
 package engine.physics;
 
 import org.joml.FrustumIntersection;
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 public interface ICollider {
 
-    Vector3f getPosition();
+    Vector3d getPosition();
     Type getType();
     boolean testFrustum(FrustumIntersection frumstum);
